Remove commands and commandsTree global variables.
